solved

+++ exercises/practice/armstrong-numbers/armstrong_numbers.zig
@@ -4,11 +4,11 @@ const pow = std.math.pow;
 const allocator = std.heap.page_allocator;
 
 pub fn isArmstrongNumber(num: u128) bool {
-    var sum: usize = 0;
+    var sum: u128 = 0;
     var str = fmt.allocPrint(allocator, "{d}", .{num}) catch return false;
 
     for (str) |c| {
-        sum += pow(usize, c - '0', str.len);
+        sum += pow(u128, c - '0', str.len);
     }
 
     return sum == num;
